Updating old tests. Updated figshare filenames so tests can check against current data
